% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{average_case_ranking}
\alias{average_case_ranking}
\title{Average case ranking.}
\usage{
average_case_ranking(p, comp_cores = 1)
}
\arguments{
\item{p}{a matrix containing colwise scoring lists to be ranked.}

\item{comp_cores}{number of cores to use for parallel processing}
}
\value{
A matrix similar to input but storing the ranks.
}
\description{
This function ranks a numeric (scoring) matrix colwise
assigning the average case rank.
}
\details{
Decreasing order - high scores appear at low rank.
If e.g. the first two candidates of a list have the same score
the rank is the mean of all involved scores, so in this case (1+2)/2 = 1.5.
}
\author{
Sarah Scharfenberg
}

