% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate Scores from sample table.}
\usage{
calculate_score(p, class_table, samples, burnin_samples = NULL, mode = NULL,
  comp_cores = NULL)
}
\arguments{
\item{p}{named Likelihood matrix}

\item{class_table}{a matrix containing as much rows as p has columns
and a column for each sample storing a complete assignemtn of all
unknown variables}

\item{samples}{the number of sampels to consider for calculation}

\item{mode}{one of "mul" or "add" or nothing if the resulting score should
be multiplied with or added to p}

\item{comp_cores}{number of cores to use for parallel processing}

\item{number}{of burn in samples to discard}
}
\value{
A matrix storing the new scores.
}
\description{
This function calculates the mean posterior estimator
given a set of samples.
}
\author{
Sarah Scharfenberg
}

