% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_synthetic_dataset.R
\name{create_synthetic_dataset_kegg}
\alias{create_synthetic_dataset_kegg}
\title{Create a synthetic dataset based on KEGG IDs.}
\usage{
create_synthetic_dataset_kegg(x, input = c("mass", "id"),
  mode = c("likelihood", "value"), value = 0.5)
}
\arguments{
\item{x}{is an vector of masses that can be named or a vector of KEGG ids.}

\item{input}{a character string defining the content of x. It can be
either \"mass\" or \"id\".}

\item{mode}{a character string defining the values of the matrix. It can
be \"likelihood\" or \"value\". (see Details)}
}
\value{
A CxM matrix containing a value for each candidate of each mass
and 0 else.
}
\description{
This method was written to create different semi-synthetic
dataset for validation purposes.
}
\details{
For each mass / id a compound search using keggFind is done and
each match is stored with a value. If likelihood was chosen a probability
is calculated using an error function taking the mean and standard
deviation of the sample into account. If value was chosen each candidate
having a mass in a fix range gets the same value.
}
\author{
Sarah Scharfenberg
}

