% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_synthetic_dataset.R
\name{eliminate_redundancies}
\alias{eliminate_redundancies}
\title{Eliminate redundancies.}
\usage{
eliminate_redundancies(p)
}
\arguments{
\item{p}{a matrix}
}
\value{
the reduced matrix.
}
\description{
This function matchs redundand rows into one row.
}

