% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_mass.R
\name{get_keggmass_by_formula}
\alias{get_keggmass_by_formula}
\title{Get mass from kegg entry by molecular formula.}
\usage{
get_keggmass_by_formula(formulas)
}
\arguments{
\item{formulas}{a vector of molecular formulas as character strings.}
}
\value{
A named vector storing the exact mass foreach formula.
}
\description{
This function searches KEGG by the given formula and returns
the exact mass for each request in a named vector.
}
\author{
Sarah Scharfenberg
}

