% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_mass.R
\name{get_keggmass_by_id}
\alias{get_keggmass_by_id}
\title{Get mass from kegg entry by kegg-id.}
\usage{
get_keggmass_by_id(ids)
}
\arguments{
\item{ids}{a vector of KEGG ids as character strings.}
}
\value{
A named vector storing the exact mass foreach id.
}
\description{
This function searches KEGG by the given ids and returns
the exact mass for each request in a named vector.
}
\author{
Sarah Scharfenberg
}

