% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{get_prediction_table}
\alias{get_prediction_table}
\title{Get prediction table.}
\usage{
get_prediction_table(p, ranking = "worst case", max_rank = 10)
}
\arguments{
\item{p}{named Likelihood matrix}

\item{ranking}{a character string specifying the rank to use.
until now only "worst case" possible.}

\item{max_rank}{the maximal rank that should be diplayed}
}
\value{
A data.frame storing 'query', 'rank', 'score', 'candidate', 'cname'
foreach compound with rank <=max_rank.
}
\description{
A function to get a human readable output of a prediction.
}
\author{
Sarah Scharfenberg
}

