% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_pubchemID_from_keggID.R
\name{get_pubchemID_from_keggID}
\alias{get_pubchemID_from_keggID}
\title{Recieve PubChem ID.}
\usage{
get_pubchemID_from_keggID(keggIDs)
}
\arguments{
\item{keggIDs}{a vector of character strings containing the KEGG ids}
}
\value{
A matrix with header keggID, SID, CID with one row foreach
input id storing the matched PubChem ids.
}
\description{
This function takes a list of KEGG Ids and tries to
convert them into PubChem IDs using the package KEGGREST.
}
\details{
As in KEGG only the SID is stored the match to CID is done
using the additional package rpubchem.
Access via internet. Be careful, long input lists might overacces
the KEGG server.
}
\author{
Sarah Scharfenberg
}

