% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_relationmatrix.R
\name{get_relationmatrix_distance}
\alias{get_relationmatrix_distance}
\title{Get relationmatrix based on chemical similarity.}
\usage{
get_relationmatrix_distance(p, distance = "tanimoto", mode = "relation",
  threshold = 0.7)
}
\arguments{
\item{p}{a matrix rownamed by CID}

\item{distance}{a character string specifying the distance that should be
calculated. Default is tanimoto.}

\item{mode}{a character string specifying the matrix value. If dist was
chosen the matrix contains the distances foreach pair. If relation was
chosen the matrix contains 1 if the distance was greater than the given
threshold and 0 else.}

\item{threshold}{a double value specifying the threshold for a chemical
similarity representing a relation. (see mode)}
}
\value{
A matrix CxC containing the specified output. (see mode)
Similarities are not directed and do not exist for a compound with itself.
}
\description{
This function creates a relationmatrix based on the chemical
similarity.
}

