% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_relationmatrix.R
\name{get_relationmatrix_distance_mem}
\alias{get_relationmatrix_distance_mem}
\title{Get relationmatrix based on chemical similarity.}
\usage{
get_relationmatrix_distance_mem(row_ids, col_ids, dist = "tanimoto",
  mode = "relation", threshold = 0.7, folder_PubChemID, verbose = FALSE,
  comp_cores = NULL, all = FALSE, fingerprints = NULL)
}
\arguments{
\item{row_ids}{a vector containing the rownames of w to be calculated by CID.}

\item{col_ids}{a vector containing the colnames of w to be calculated  by CID.}

\item{dist}{a character string specifying the distance that should be
calculated. Default is tanimoto.}

\item{mode}{a character string specifying the matrix value. If dist was
chosen the matrix contains the distances foreach pair. If relation was
chosen the matrix contains 1 if the distance was greater than the given
threshold and 0 else.}

\item{threshold}{a double value specifying the threshold for a chemical
similarity representing a relation. (see mode)}

\item{folder_PubChemID}{a character string specifying the location of a
prepared local pubchem mirror}

\item{verbose}{switch detailed output on}

\item{comp_cores}{number of cores to use for parallel processing}

\item{all}{a logical denoting wheter the complete matrix
should be calculated. If FALSE the upper trinangular matrix
is calculated and returned as symmetric matrix.}
}
\value{
A matrix CxC containing the specified output. (see mode)
Similarities are not directed and do not exist for a compound with itself.
Careful with big C and similarity output this
is no sparse matrix!
}
\description{
This function creates a relationmatrix based on
their chemical similarity. It was desinged for real big datasets
(like >1Mio candidates) as they could run out of memory. Therefore
it is slower than get_relationmatrix_distance_par
}

