% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_relationmatrix.R
\name{get_relationmatrix_rp}
\alias{get_relationmatrix_rp}
\title{Get relationmatrix from reaction pairs.}
\usage{
get_relationmatrix_rp(p, rp)
}
\arguments{
\item{p}{a rownamed matrix containing the likelihoods (dimension CxM)}

\item{rp}{a data.table containing reaction pairs in colum V1 and V2,
respectively.}
}
\value{
A sparse matrix CxC containing 1 if there is a reaction pair containing
these compounds, 0 else. Connections are not directed and do not exist
for a compound with itself.
}
\description{
This function creates a relationmatrix from a list of
reaction pairs.
}
\details{
A list of reaction pairs can be retrieved for example from
the kegg reactions database.
The ids of rp should match the ids of p. If not try to translate them
with ...
}
\author{
Sarah Scharfenberg
}

