% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_relationmatrix.R
\name{get_relationmatrix_zero}
\alias{get_relationmatrix_zero}
\title{Get relationmatrix 0.}
\usage{
get_relationmatrix_zero(p)
}
\arguments{
\item{p}{a rownamed matrix containing the likelihoods (dimension CxM)}
}
\value{
A matrix CxC containing 0.
}
\description{
This function creates a relationmatrix without
any connection. It was written to validate the influence of
different connections.
}
\author{
Sarah Scharfenberg
}

