% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_true.R
\name{get_true}
\alias{get_true}
\title{Get true assignment.}
\usage{
get_true(p)
}
\arguments{
\item{p}{a row- and colnamed matrix.}
}
\value{
A matrix of the same size as p containing 1 if colname equals
rowname and 0 else.
}
\description{
For testing only.
}
\author{
Sarah Scharfenberg
}

