\name{GlobMetab-package}
\alias{GlobMetab-package}
\alias{GlobMetab}
\docType{package}
\title{
A package for global identification of metabolites.
}
\description{
The package is supposed to help identifying 
  metabolites of a biological sample taking into account 
    that compounds of the same sample are in some way related to eachother.
}
\details{
\tabular{ll}{
Package: \tab GlobMetab\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-11\cr
License: \tab GPL-3\cr
}
Hier folgt noch eine detaillierte Anweisung zur Nutzung des Pakets.
}
\author{
Sarah Scharfenberg

Maintainer: Sarah Scharfenberg <sscharfe@ipb-halle.de>
}
\references{
 # Literature or other references for background information ~~
}
 # Optionally other standard keywords, ~~
 # one per line, from file KEYWORDS in ~~
 # the R documentation directory ~~
\keyword{ package }
\seealso{
 #Optional links to other man pages, e.g. ~~
 # \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
 # simple examples of the most important functions ~~
}
