% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare a dataset.}
\usage{
prepare_dataset(file)
}
\arguments{
\item{file}{a string containing the path to a folder storing a set of
sdf files.}
}
\value{
p a CxM matrix where m is the number of queries read and C is
the number of possible compounds of all candidate lists.
}
\description{
A method to prepare a whole set of sdf files that
are supposed to be processed together.
}
\author{
Sarah Scharfenberg
}

