% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.sdf.R
\name{read.sdf_sparse}
\alias{read.sdf_sparse}
\title{A function to read a long list of sdf files parallel into
a sparse matrix p.}
\usage{
read.sdf_sparse(path, id = "DatabaseID", add = 0.01,
  filter_by_formula = FALSE, formulas = NULL, filter_by_peaks = FALSE,
  npe = NULL, NumberPeaksExplained = NULL, formula_by_SMILES = FALSE,
  verbose = FALSE, comp_cores = 1)
}
\arguments{
\item{path}{a character string specifying the absolute path of
the folder containing the sdf files to be read.}

\item{id}{a character string identifying the property the
compounds should identified by eg DatabaseID, Identifyer or
InChIKey1}

\item{add}{a numeric value that should be added to each input score
to distinguish a scroe of 0 from a candidate not appaering in the
candidate list at all.}

\item{filter_by_formula}{a logical indicating whether the canidate
lists should be cut to only those having the same molecular formula}

\item{formulas}{a named character vector specifying a formula for
each filename in the folder. If not provided when filter_by_formula
is TRUE then the corresponding list is not filtered.}

\item{filter_by_peaks}{a logical indicating whether the candidate
lists should be cut to only those explaining at least a given
number of peaks.}

\item{npe}{a character string defining the sdf property where the
number of explained peaks can be found}

\item{NumberPeaksExplained}{the number of peaks that should be at least
explained by the MetFrag-candidates peakspectrum}

\item{formula_by_SMILES}{a logical indicating whether the molecular
formula for the candidates should be calculated using the SMILES
property rather than the MolecularFormula, e.g. if MolecularFormula
is not provided.}

\item{verbose}{logical indicating if a detailed output is wanted}

\item{comp_cores}{number of cores to use for parallel processing
#}
}
\description{
A function to read a long list of sdf files parallel into
a sparse matrix p.
}

