% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{read_sdf}
\alias{read_sdf}
\title{Read sdf file.}
\usage{
read_sdf(file)
}
\arguments{
\item{file}{a string containing the path to a sdf file.}
}
\value{
...
}
\description{
A method to read a MetFrag result list from an sdf file.
}
\author{
Sarah Scharfenberg
}

