% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_relationmatrix.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity of 2 smiles.}
\usage{
similarity(s1, s2, dist = "tanimoto")
}
\arguments{
\item{s1}{smiles string of compound 1}

\item{s2}{smiles string of compound 2}

\item{dist}{the distance to calculate as parameter method defined
in rcdk::distance}
}
\value{
The distance of the two smiles.
}
\description{
This function calculates the chosen distance between two
smiles strings.
}
\author{
Sarah Scharfenberg
}

