% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_matrix}
\alias{validate_matrix}
\title{Validate prediction.}
\usage{
validate_matrix(true, pred)
}
\arguments{
\item{true}{a matrix containing the correct assignment.}

\item{pred}{a matrix containing a predicted assignment.}
}
\value{
A list storing the percentage of correct assigned and false
assigned compounds.
#
}
\description{
This function checks colwise equality of two assignments z.
E.g. a predicted and the true or changes between different predictions.
}
\author{
Sarah Scharfenberg
}

