% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_rank}
\alias{validate_rank}
\title{Validation of two rankings.}
\usage{
validate_rank(z_true, p, p_new)
}
\arguments{
\item{z_true}{a named numeric vector containg forach query (named) the
index of the correct compound regarding the matrix p}

\item{p}{a named matrix containing the intial ranking}

\item{p_new}{a named matrix containing a predicted ranking}
}
\value{
A list storing
$'better rank' - number and percentage of queries having a better rank
$'worse rank' - number and percentage of queries having a worse rank
$'same rank' - number and percentage of queries having the same rank
$'true id lost' - number and percent of ids lost because true id unknown
}
\description{
This function validates a rank given a vector
containing the true compund ids.
}
\author{
Sarah Scharfenberg
}

