% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_vector}
\alias{validate_vector}
\title{Validate prediction.}
\usage{
validate_vector(true, pred)
}
\arguments{
\item{true}{a vector containing the correct assignment.}

\item{pred}{a vector containing a predicted assignment.}
}
\value{
A list storing the percentage of correct assigned and false
assigned compounds.
#
}
\description{
This function checks equality of a two assignments z
reduced to the index of each assignment.
E.g. a predicted and the true or changes between different predictions.
}
\author{
Sarah Scharfenberg
}

