% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{worst_case_ranking}
\alias{worst_case_ranking}
\title{Worst case ranking.}
\usage{
worst_case_ranking(p)
}
\arguments{
\item{p}{a matrix containing colwise scoring lists to be ranked.}
}
\value{
A matrix similar to input but storing the ranks.
}
\description{
This function ranks a numeric (scoring) matrix colwise
assigning the worst case rank.
}
\details{
Decreasing order - high scores appear at low rank.
If e.g. the first two candidates of a list have the same score
the rank is the worst possible, so in this case 2 for both.
}
\author{
Sarah Scharfenberg
}

