% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/validate.R
\name{calculate_score}
\alias{calculate_score}
\title{Calculate Scores from sample table.}
\usage{
calculate_score(p, class_table, samples, mode = NULL, last = TRUE)
}
\arguments{
\item{p}{named Likelihood matrix}

\item{class_table}{a matrix containing as much rows as p has columns
and a column for each sample storing a complete assignemtn of all
unknown variables}

\item{samples}{the number of sampels to consider for calculation}

\item{mode}{one of "mul" or "add" or nothing if the resulting score should
be multiplied with or added to p}

\item{last}{a logical value indicating whether the samples should be
obtained from the end (e.g. the last 1000 samples, equivalent to a
burn in phase of nrow(class_table)-samples)}
}
\value{
A matrix storing the new scores.
}
\description{
This function calculates the mean posterior estimator
given a set of samples.
}
\author{
Sarah Scharfenberg
}

