% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.sdf.R
\name{eliminate_redundancies}
\alias{eliminate_redundancies}
\title{Eliminate redundancies.}
\usage{
eliminate_redundancies(p, log = NULL)
}
\arguments{
\item{p}{a matrix}

\item{log}{a character String specifying the path to a pubchem database
log file, containing information about the latest update, omitting
ids greater that the last id specified there}
}
\value{
the reduced matrix.
}
\description{
This function matchs redundand rows into one row.
}

