% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_relationmatrix.R
\name{similarity}
\alias{similarity}
\title{Calculate similarity of 2 smiles.}
\usage{
similarity(fp1, fp2, dist = "tanimoto")
}
\arguments{
\item{fp1}{fingerprint of compound 1}

\item{fp2}{fingerprint of compound 2}

\item{dist}{the distance to calculate as parameter method defined
in rcdk::distance}
}
\value{
The distance of the two fringerprints.
}
\description{
This function calculates the chosen distance between two
rcdk smiles.
}
\author{
Sarah Scharfenberg
}

